+++ src/validators/numeric/index.ts
@@ -1,2 +1,4 @@
 export { IsInteger } from "./is-integer";
 export { IsFinite } from "./is-finite";
+
+export { IsPositive } from "./is-positive";