+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [ZeroVer](http://0ver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Validator `IsPositive`.
+
 ## [0.1.5] - 2018-04-17
 
 ### Added